+++ doc/changelog.rst
@@ -21,6 +21,9 @@ Contributors
 
 Added
 -----
+- Add three EBSD patterns from the same position of a single crystal Si sample
+  with varying sample-screen distances to the data module via external repo.
+  (`#318 <https://github.com/pyxem/kikuchipy/pull/318>`_)
 - Reading of NORDIF calibration patterns specified in a setting file into an
   EBSD signal. (`#317 <https://github.com/pyxem/kikuchipy/pull/317>`_)